+++ abe/gpsw_test.go
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2018 XLAB d.o.o
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package abe_test
 
 import (
@@ -10,11 +26,11 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestABE(t *testing.T) {
+func TestGSPW(t *testing.T) {
 	// create a new ABE struct with the universe of l possible
 	// attributes (attributes are denoted by the integers in [0, l)
 	l := 10
-	a := abe.NewABE(l)
+	a := abe.NewGSPW(l)
 
 	// generate a public key and a secret key for the scheme
 	pubKey, sk, err := a.GenerateMasterKeys()
